Refactor for better project structure
