+++ src/main/java/kat/chatbot/TaskStorage.java
@@ -1,6 +1,8 @@
-package chatbot;
+package kat.chatbot;
 
 import java.util.List;
+
+import kat.tasks.Task;
 
 /**
  * Represents storage system for managing tasks.
