+++ src/main/resources/view/MainWindow.fxml
@@ -11,7 +11,7 @@
             prefHeight="600.0" prefWidth="400.0"
             xmlns="http://javafx.com/javafx/17"
             xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="chatbot.ui.MainWindow">
+            fx:controller="kat.ui.MainWindow">
     <children>
         <VBox AnchorPane.topAnchor="0.0"
               AnchorPane.leftAnchor="0.0"
@@ -40,8 +40,7 @@
                            prefHeight="41.0"
                            prefWidth="324.0"
                            AnchorPane.bottomAnchor="1.0"/>
-                <Button fx:id="sendButton"
-                        layoutX="324.0"
+                <Button layoutX="324.0"
                         layoutY="558.0"
                         mnemonicParsing="false"
                         onAction="#handleUserInput"
